+++ RUFAS/biophysical/manure/storage/storage.py
@@ -6,11 +6,24 @@ from RUFAS.biophysical.manure.processor import Processor
 from RUFAS.current_day_conditions import CurrentDayConditions
 from RUFAS.data_structures.animal_to_manure_connection import ManureStream
 from RUFAS.general_constants import GeneralConstants
+from RUFAS.input_manager import InputManager
 from RUFAS.rufas_time import RufasTime
 from RUFAS.util import Utility
 
 from .storage_cover import StorageCover
 from ..manure_constants import ManureConstants
+
+MANURE_CONVERSION_CONSTANT = 0.1175
+"""Factor to estimate m^3 of herd-wide manure produced per day per mature cow housed on the farm (m^3/day)."""
+
+FREEBOARD_CONSTANT = 1.20
+"""Represents 20% volume allowance above the maximum volume of a slurry or liquid manure storage (unitless)."""
+
+DEPTH_CONSTANT = 4.572
+"""Value for slurry or liquid manure storage depth (m)."""
+
+PRECIPITATION_CONSTANT = 0.25
+"""The annual precipitation constant value (m)."""
 
 
 class Storage(Processor):
@@ -65,11 +78,29 @@ class Storage(Processor):
         self._storage_time_period = storage_time_period
         self._surface_area = surface_area
         self._manure_to_process = ManureStream.make_empty_manure_stream()
+        self.__post_init__()
+
+    def __post_init__(self) -> None:
+        """
+        Post-initialization method to calculate the surface area of the storage.
+        """
+        if self._surface_area is None:
+            self._calculate_surface_area()
 
     @property
     def is_overflowing(self) -> bool:
         """True if the manure in storage exceeds the storage's volumetric capacity, else False."""
         return self._stored_manure.volume > self._capacity
+
+    def _calculate_surface_area(self) -> None:
+        """
+        Calculates the surface area of the storage.
+        """
+        cow_num = InputManager().get_data("animal.herd_information.cow_num")
+        self._surface_area = (
+            (cow_num * MANURE_CONVERSION_CONSTANT * self._storage_time_period * FREEBOARD_CONSTANT)
+            / (DEPTH_CONSTANT - PRECIPITATION_CONSTANT)
+        )
 
     def receive_manure(self, manure: ManureStream) -> None:
         """Receives manure and puts it in storage to be processed."""